requirements.txt
psycopg2-binary>=2.9.0
python-jobspy>=1.1.80
selenium>=4.18.1
undetected-chromedriver>=3.5.5
beautifulsoup4>=4.12.3
requests>=2.31.0
schedule>=1.2.1
tabulate>=0.9.0
pandas>=2.0.0
lxml>=5.0.0
flask>=3.0.0
flask-login>=0.6.3
werkzeug>=3.0.0
gunicorn>=21.0.0
